+++ Drivers/lora/lora_drv.h
@@ -27,7 +27,7 @@ typedef struct xLoRaIf_t {
 extern LoRaIf_t LoRaInterface;
 
 bool lora_init(void);
-bool lora_proc_payload(uint8_t* rx_payload, uint8_t rx_size);
+bool lora_proc_payload(uint8_t* rx_payload, uint32_t rx_size);
 bool lora_send_queue(uint8_t* const tx_payload, uint32_t len);
 bool lora_process(void);
